# swagger.yaml
openapi: 3.0.0
info:
  title: Flex Living Reviews API
  version: 1.0.0
  description: |
    API for managing property reviews and dashboard functionality.
    
    This API provides endpoints for:
    - Fetching and normalizing Hostaway reviews (Assessment requirement)
    - Managing reviews with filtering and pagination
    - Admin operations for review approval
    - Google Reviews integration
    - Health monitoring
  contact:
    name: Flex Living
    email: dev@flexliving.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:3000
    description: Memory Server (Development)
  - url: http://localhost:4000
    description: PostgreSQL Server (Production)

security:
  - AdminKey: []

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Returns server health status and data counts
      tags:
        - Health
      security: []
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Memory server is running!"
                  data:
                    type: object
                    properties:
                      reviews:
                        type: number
                        description: Number of reviews in memory
                      listings:
                        type: number
                        description: Number of listings in memory

  /api/reviews/hostaway:
    get:
      summary: Get Hostaway reviews (Assessment endpoint)
      description: |
        **REQUIRED BY THE ASSESSMENT**
        
        Fetches and returns normalized Hostaway review data with pagination.
        This endpoint normalizes the Hostaway API response format and provides
        structured, usable data for the frontend dashboard.
      tags:
        - Reviews
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          description: Number of reviews per page
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: listing
          in: query
          description: Filter by listing name
          required: false
          schema:
            type: string
        - name: rating
          in: query
          description: Filter by rating (1-5 stars)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: status
          in: query
          description: Filter by review status
          required: false
          schema:
            type: string
            enum: [published, hidden, pending]
        - name: approved
          in: query
          description: Filter by approval status
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of Hostaway reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/reviews:
    get:
      summary: Get all reviews with filtering
      description: |
        Fetches reviews from all sources (Hostaway, Google) with advanced filtering options.
        Supports pagination, search, and multiple filter criteria.
      tags:
        - Reviews
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: pageSize
          in: query
          description: Number of reviews per page
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: approved
          in: query
          description: Filter by approval status
          required: false
          schema:
            type: boolean
        - name: channel
          in: query
          description: Filter by review channel
          required: false
          schema:
            type: string
            enum: [hostaway, google]
        - name: q
          in: query
          description: Search in review comments
          required: false
          schema:
            type: string
        - name: rating
          in: query
          description: Filter by rating (1-5 stars)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: listing
          in: query
          description: Filter by listing name
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of filtered reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/admin/reviews/{id}/approve:
    patch:
      summary: Approve a review for public display
      description: |
        Approves a review to be displayed on the public website.
        Requires admin authentication via X-Admin-Key header.
      tags:
        - Admin
      parameters:
        - name: id
          in: path
          description: Review ID to approve
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Review approved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Review approved successfully"
                  review:
                    $ref: '#/components/schemas/Review'
        '400':
          description: Bad request - Invalid review ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - Invalid admin key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/google/seed:
    post:
      summary: Seed Google reviews
      description: |
        Populates the system with mock Google reviews for testing and development.
        This endpoint creates sample Google reviews to demonstrate the integration.
      tags:
        - Google Reviews
      responses:
        '200':
          description: Google reviews seeded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Google reviews seeded successfully"
                  count:
                    type: integer
                    description: Number of reviews added
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/google/places/{placeId}:
    get:
      summary: Fetch reviews from Google Places API
      description: |
        Fetches reviews from Google Places API for a specific place.
        Currently returns mock data, but can be extended for real Google Places integration.
      tags:
        - Google Reviews
      parameters:
        - name: placeId
          in: path
          description: Google Places ID
          required: true
          schema:
            type: string
            example: "ChIJN1t_tDeuEmsRUsoyG83frY4"
      responses:
        '200':
          description: Google reviews fetched successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  placeId:
                    type: string
                    description: The Google Places ID that was queried
        '400':
          description: Invalid place ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin key for accessing protected endpoints. Use 'admin-secret-key-123' for testing.

  schemas:
    Review:
      type: object
      required:
        - id
        - listingId
        - rating
        - channel
        - submittedAt
      properties:
        id:
          type: string
          description: Unique identifier for the review
          example: "hostaway:7453"
        listingId:
          type: string
          description: ID of the property listing
          example: "listing-1"
        listingName:
          type: string
          description: Name of the property
          example: "2B N1 A - 29 Shoreditch Heights"
        guestName:
          type: string
          description: Name of the guest who left the review
          example: "Shane Finkelstein"
        rating:
          type: number
          minimum: 1
          maximum: 5
          description: Overall rating (1-5 stars)
          example: 5
        comment:
          type: string
          description: Review comment text
          example: "Shane and family are wonderful! Would definitely host again :)"
        type:
          type: string
          enum: [host-to-guest, public]
          description: Type of review
          example: "host-to-guest"
        status:
          type: string
          enum: [published, hidden, pending]
          description: Review status
          example: "published"
        channel:
          type: string
          enum: [hostaway, google]
          description: Review source channel
          example: "hostaway"
        submittedAt:
          type: string
          format: date-time
          description: When the review was submitted
          example: "2020-08-21T22:45:14.000Z"
        approved:
          type: boolean
          description: Whether the review is approved for public display
          example: false
        categories:
          type: array
          description: Category-specific ratings
          items:
            type: object
            properties:
              category:
                type: string
                enum: [cleanliness, communication, respect_house_rules, location, value]
                example: "cleanliness"
              rating:
                type: number
                minimum: 1
                maximum: 5
                example: 5

    Listing:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: Unique identifier for the listing
          example: "listing-1"
        name:
          type: string
          description: Name of the property
          example: "2B N1 A - 29 Shoreditch Heights"
        address:
          type: string
          description: Property address
          example: "29 Shoreditch Heights, London"
        averageRating:
          type: number
          description: Average rating across all reviews
          example: 4.5
        totalReviews:
          type: number
          description: Total number of reviews
          example: 12

    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of items per page
          example: 50
        total:
          type: integer
          description: Total number of items
          example: 150
        totalPages:
          type: integer
          description: Total number of pages
          example: 3

    Error:
      type: object
      required:
        - error
        - status
      properties:
        error:
          type: string
          description: Error message
          example: "Review not found"
        status:
          type: number
          description: HTTP status code
          example: 404
        details:
          type: string
          description: Additional error details
          example: "The review with ID 'invalid-id' does not exist"

tags:
  - name: Health
    description: Health monitoring endpoints
  - name: Reviews
    description: Review management and retrieval
  - name: Admin
    description: Administrative operations requiring authentication
  - name: Google Reviews
    description: Google Reviews integration and management